Refactored code for unit tests
